Fix stuff
Add build script for Unix.
Rename files to lowercase.
Add platform-agnostic sleep command.
