+++ build.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+g++ -o main main.cpp gameManager.cpp includes.cpp Objects/character.cpp Objects/enemy.cpp Objects/mainChar.cpp
