+++ includes.cpp
@@ -1,11 +1,11 @@
 #include <iostream> // Reading 'help' file
 #include <cstdlib> // rand(), srand()
-#include <windows.h> //Sleep()
 #include <fstream> // ifstream
 
-#include "Objects/MainChar.h"
-#include "Objects/Enemy.h"
+#include "Objects/mainChar.h"
+#include "Objects/enemy.h"
 #include "includes.h"
+#include "util.h"
 
 // ----- Defining Non-member functions ----- //
 
@@ -21,7 +21,7 @@ int dead_check(T& thing)
 {
     if(thing.hp<=0)
     {
-        Sleep(1000);
+        tba_sleep(1000);
         std::cout << "The " << thing.name << " has been slaughtered." << std::endl << std::endl;
         return 1;
     }
@@ -36,9 +36,9 @@ int dead_check<>(MainChar& player)
 {
     if( player.hp <= 0 )
     {
-        Sleep(1000);
+        tba_sleep(1000);
         std::cout << "You have died!" << std::endl;
-        Sleep(1000);
+        tba_sleep(1000);
         return 1;
     }
     else
@@ -73,7 +73,7 @@ void battle(MainChar& player, Enemy& foe)
     while( ( foe.hp > 0 ) && ( player.hp > 0 ) )
     {
         //--- Enemy turn ---//
-        Sleep(1000);
+        tba_sleep(1000);
         foeChoice = random( spellProb(foe) );
         if( (foeChoice == 0) )  //if the enemy will cast a spell
         {
@@ -88,15 +88,15 @@ void battle(MainChar& player, Enemy& foe)
 
         //--- Player turn ----//
 
-        Sleep(1000);
+        tba_sleep(1000);
         if(player.hp>0)
         {
             // UI text
-            std::cout << player.name << ": hit points " << player.hp << "/att. power " << player.att_pow << "/magic points " << player.magic << endl;
+            std::cout << player.name << ": hit points " << player.hp << "/att. power " << player.att_pow << "/magic points " << player.magic << std::endl;
             std::cout << "The " << foe.name << " has " << foe.hp << " hit points remaining." << std::endl << std::endl;
-            Sleep(1500);
+            tba_sleep(1500);
             std::cout << "What will you do?" << std::endl;
-            Sleep(500);
+            tba_sleep(500);
             std::cout << "Choose: 'attack', 'bolt' or 'aura'." << std::endl;
             std::cout << "Type 'help' if you need the rules of combat explained." << std::endl;
             std::string turn_choice = "";
