+++ main.cpp
@@ -1,10 +1,10 @@
-#include "GameManager.h"
+#include "gameManager.h"
 
 // All we need to do here is call the GameManager constructor, 
 //  since this will execute the game loop 
 
 int main()
 {
-    GameManager *game = new GameManager(); 
+    gameManager *game = new gameManager(); 
     return 0;
 }
